First Implementation of Boxing System
Two new classe:
 - Box, the elementary container
 - BoxSet, the set of Boxes that will be called by System

So far, the BoxSet generates and updates the Boxes neighboroods
correctly, using periodic boundary consitions defined in
System class.
Boxes do not include any container yet, and thus system cannot be used so far.
Next step will be to introduce a container and allow System to put some
particles in it. A box should then be able to give to a particle the
complete set of relevant particles in its vicinity.