First Implementation of polydispersity.
Should not use yet with particle with radius!=1 :)

Polydispersity taken into account into:
 * Simulation input
 * Contact interactions, for the force computation part
 * Lubrication matrix, and strain rate term

Polydispersity is still to be added into:
 * Boxing algorithm ( it is essential! )
 * Stress computation
 * Configuration generator

As a side effect, I redefined normal vector for interactions as
the opposite of the previous definition.
SO FAR, THIS (and not polydispersity itself) INTRODUCES A BUG:
PARTICLES DO OVERLAP

It is unclear to me so far where I messed up.