Implementing Incomplete Cholesky Preconditioner
An Incomplete Cholesky Preconditioner is added for the iterative solver
in StokesSolver class.
It is based on Ifpack package of Trilinos.
Right now, it is not activated, but it can be switched on by
decommenting the corresponding line in complete_RFU() method.

Performance is poor at this stage, so parameter tuning is
probably required.