Removing references to dimensionless_number in the System class.
Now this is no more meaningful, as many forces can be used concurrently,
hence there's many dimensionless numbers.

dimensionless_number was used to scale forces when the hydrodynamic
units of forces are used. Now this scaling is done in the Simulation
class at initialization.

Note:
It was useful to rescale these forces after
init when the hydro units were used with variable shear rate, like in
the stress controlled case.
Now we prefer NOT using the hydro scale in this case, so there should
not be any need to rescale the forces after init. This should be avoided.