New Makefile mechanism, big change
I changed the mechanism for environment options in the Makefile. Many
different options were starting to accumulate here. Moreover every
local change needed to be taken care of with git, which is annoying.

Now, environment dependent variables are given in
config/Makefile_config.mk . This file is NOT commited however, as it
is all about local variables. Instead there is template file
Makefile_config_generic.mk, that you can specialize and copy to
Makefile_config.mk before compiling.

The old Makefile is kept and commited in Makefile.old